Initial portion of the JavaScript canvas tutorial...
We start with drawing a circle on a canvas element. The kids construct an HTML
document with just a `canvas` and `script` element in the body. They add CSS
that puts a red border around the canvas, so that we can see where it is on the
page.

In the script, add variables for the canvas and the drawing context. Explain
what variables are. Hopefully someone in the class will have some pre-algebraic
experience in school at least and will be familiar with the concept of a
variable in math. While it's not the same, it can be a good jumping-off point.

The students type out the functions (dictation or copy-paste may be the best
way, as some of these functions have no basis yet). They should have been
introduced to functions before now, maybe with alerts or something.

Draw the arc. If they've had geometry, they might know what it is. If not, you
may have some explaining to do.