Draw another circle to the right of the first
Get the kids to copy the code that drew their first circle in order to draw
another one to the right of it. Point out the similarity of the two blocks and
how they really only differ by a couple numbers. Ask them what they'd have to
do in order to draw a third circle.