Use a function to cut down on code repeat
Explain to them that they can use a function to cut down on the amount of code
that they have to write to achieve any particular task more than once. Have
them open up the developer console in their browser and draw a few more circles
on the fly using their new function.